move versioning into internal/
There is no need to expose versioning and let it be imported outside of
the oracle code
