+++ Makefile
@@ -4,8 +4,8 @@ print-version:
 	@echo "Version: ${VERSION}"
 
 build:
-	@go build -v -ldflags "-X github.com/masa-finance/masa-oracle/pkg/config.ApplicationVersion=${VERSION}" -o ./bin/masa-node ./cmd/masa-node
-	@go build -v -ldflags "-X github.com/masa-finance/masa-oracle/pkg/config.ApplicationVersion=${VERSION}" -o ./bin/masa-node-cli ./cmd/masa-node-cli
+	@go build -v -ldflags "-X github.com/masa-finance/masa-oracle/internal/versioning.ApplicationVersion=${VERSION}" -o ./bin/masa-node ./cmd/masa-node
+	@go build -v -ldflags "-X github.com/masa-finance/masa-oracle/internal/versioning.ApplicationVersion=${VERSION}" -o ./bin/masa-node-cli ./cmd/masa-node-cli
 	
 install:
 	@sh ./node_install.sh
