+++ cmd/masa-node-cli/components.go
@@ -5,7 +5,8 @@ import (
 	"strings"
 
 	"github.com/gdamore/tcell/v2"
-	"github.com/masa-finance/masa-oracle/pkg/config"
+	"github.com/masa-finance/masa-oracle/internal/versioning"
+
 	"github.com/rivo/tview"
 )
 
@@ -125,7 +126,7 @@ const (
 	navigation = `[yellow]use keys or mouse to navigate`
 )
 
-var version string = fmt.Sprintf(`[green]Application version: %s\n[green]Protocol Version: %s`, config.ApplicationVersion, config.ProtocolVersion)
+var version string = fmt.Sprintf(`[green]Application version: %s\n[green]Protocol Version: %s`, versioning.ApplicationVersion, versioning.ProtocolVersion)
 
 // Splash shows the app info
 func Splash() (content tview.Primitive) {
