+++ internal/versioning/version.go
@@ -1,4 +1,4 @@
-package config
+package versioning
 
 var (
 	ApplicationVersion string
