+++ pkg/pubsub/node_data.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/masa-finance/masa-oracle/internal/versioning"
 	"github.com/masa-finance/masa-oracle/pkg/config"
 
 	"github.com/libp2p/go-libp2p/core/host"
@@ -262,7 +263,7 @@ func GetSelfNodeDataJson(host host.Host, isStaked bool) []byte {
 		IsWebScraper:      config.GetInstance().WebScraper,
 		IsValidator:       config.GetInstance().Validator,
 		IsActive:          true,
-		Version:           config.ProtocolVersion,
+		Version:           versioning.ProtocolVersion,
 	}
 
 	// Convert NodeData to JSON
